Optimize fetch for issuers and quotes
Fetch only those issuers missing in local cache, and for quotes -
which are either missing or stale by timestamp. Every pull to refresh
will requery index from network and then reload issuers that are appeared
to be new in index and hence missing in local cache. Quotes are also
fetched from local cache first and if there is no such quote for a given
ticket or quote is stale (it's timestamp is more than a day ago) - quote
will be fetched from the network and then cached.